refactor: 클래스를 domain 과 ui 별로 정리
