+++ src/ui/InputView.js
@@ -1,5 +1,5 @@
 const { Console } = require('@woowacourse/mission-utils');
-const { MESSAGE } = require('./utils/constants');
+const { MESSAGE } = require('../utils/constants');
 
 class InputView {
   constructor() {}
